Refactor AddProductViewBodyBlocConsumer to conditionally show progress indicator
The `AddProdcutViewBodyBlocConsumer` widget now conditionally displays the `ModalProgressHUD` based on the state of the `AddProductBloc`. Previously, the progress indicator was always shown, regardless of the state. This change ensures that the progress indicator is only displayed when the `AddProductLoading` state is active, improving the user experience by preventing unnecessary visual clutter. Additionally, the `ModalProgressHUD` widget's `progressIndicator` property was updated to use a `const` constructor for better performance.